DeltaQ.Diagram.Internal

Description

INTERNAL MODULE. USE DeltaQ.Diagram INSTEAD!

Render outcome expressions as outcome diagrams.

Documentation

out :: O -> IO ()

renderOutcomeDiagram :: O -> Diagram SVG

Render an outcome expression as an outcome diagram.

type X = Int

type Y = Int

type Prob = Rational

type Description = String

data Op

Constructors

OFirst

OLast

OChoices [Prob]

data Op0

Tiles representing predefined outcomes

Constructors

ONever

OWait0

OWait Rational

data Token

Data attached to a Tile.

Constructors

VarT String

Outcome Op0

Location Description

Horizontal

Close (Maybe Description) [Y]

Open Op [Y]

data Tile

Constructors

Tile X Y Token

renderTiles :: [Tile] -> Diagram SVG

Render a collection of tiles.

renderToken :: Token -> Diagram SVG

Render a single Token associated with a Tile.

renderOp0Symbol :: Op0 -> Diagram SVG

Render the symbol that represents a known outcome.

renderOpSymbol :: Op -> Diagram SVG

Render the symbol that represents an operation with multiple arguments

showProb :: Rational -> String

Show a probability in scientific notation with two digits of precision.

data EdgeOutcome

Outcome-related data that we associate with an edge.

Constructors

Empty
No outcome is associated with this edge.

Term (Term String) (Maybe Description)
Term with maybe a description of its ending observation location.

type EdgeData = (Y, EdgeOutcome)

layout :: Term String -> [Tile]

Layout an outcome diagram.

emitColumn :: X -> Shrub EdgeData -> ([Tile], Shrub EdgeData)

Emit the next column, possible empty.

isVarOrKnown :: Term v -> Bool

Check that a given Term is a Var or known constant.

isEmpty :: EdgeData -> Bool

Check that a given EdgeData contains no outcome-related data.

expandSeq :: Shrub EdgeData -> Shrub EdgeData

Expand all Seq in the foliage as a sequence of twigs.

Merge observation locations that follow after a vertical term into the term annotation.

emitLoc :: X -> Shrub EdgeData -> ([Tile], Shrub EdgeData)

Emit a column with the next observation locations.

emitVar :: X -> Shrub EdgeData -> ([Tile], Shrub EdgeData)

Emit a column with the next named items.

emitVertical :: X -> Shrub EdgeData -> ([Tile], Shrub EdgeData)

Emit a column with the next vertical items.

isVertical :: Term v -> Bool

Outcomes that open a new vertical.

maxVertical :: Term v -> Int

Maximal number of outcomes that are shown vertically. Parallel operations are shown vertically, but also Choices.

hasGroupClose :: Shrub EdgeData -> Bool

Check whether there is a group of silent foliage that should be closed.

hasIsolatedTwigs :: Shrub EdgeData -> Bool

Check whether there are any foliage edges that have no expressions and no siblings.

foliageBushes :: Shrub a -> [Shrub a]

foliageBushes returns the collection of all subtrees that consist only of foliage. (These trees have height 1).

Trees where the immediate children of a root contain both foliage and non-foliage are not returned here!

emitGroupClose :: X -> Shrub EdgeData -> ([Tile], Shrub EdgeData)

Emit a column that closes all groups that can be closed at the moment.

Special case: If the next-higher branch is an observation location with a label, we merge that location into the close.

dropIsolatedTwigs :: Shrub EdgeData -> Shrub EdgeData

Drop all foliage edges that have no expressions and have no siblings.

isIsolatedTwig :: Shrub EdgeData -> Bool

Check whether a Shrub is a twig without expression.

newtype Shrub a

A Shrub is a rooted tree that has labels (only) on the edges (as opposed to on the leaves).

Constructors

Branch

Fields

unBranch :: [(a, Shrub a)]

isRoot :: Shrub a -> Bool

Test whether a Shrub is an isolated root.

root :: Shrub a

The Shrub that is an isolated root.

isTwig :: Shrub a -> Bool

Test whether a Shrub is a single edge.

twig :: a -> Shrub a

Construct a Shrub with a single edge.

foliage :: Shrub a -> [a]

The foliage of a Shrub is the collection of edges that end in a leaf.

This function returns the labels of the foliage of the given Shrub.

updateFoliage :: (a -> Shrub a) -> Shrub a -> Shrub a

Replace the foliage by news Shrubs.

Each edge in the foliage is replaced by the new Shrub whose root is attached to the end of the edge that is not a leaf.

Special cases:

If the new Shrub is root, the foliage edge will be removed.

If the new Shrub is a twig, the foliage edge will get a new label.

example0 :: O

example1 :: O

example2 :: O

exampleS :: O

example3 :: O

exampleCache :: O