DeltaQ.Expr

Description

The type O represents outcome expressions that can be

inspected,

rendered as outcome diagrams, and

converted to a probability distribution of completion times.

Outcome expressions

data O

Outcome expression.

This type includes graphical annotations for the rendering as outcome diagram.

var :: String -> O

Outcome variable, given by a unique name.

substitute

Arguments

:: (String -> O)
Assignment of variable names to outcome expressions.

-> O
Outcome expression in which we substitute.

-> O

Substitute all variables by outcome expressions.

toDeltaQ :: (String -> DQ) -> O -> DQ

Outcome expressions can be mapped to probability distributions of completion times DQ, provided that we know how to map var.

Outcome terms

termFromOutcome :: O -> Term String

Inspect an outcome expression O through its Term representation.

The result Term satisfies isNormalizedAssoc.

outcomeFromTerm :: Term String -> O

Construct an outcome expression O from an outcome Term.

data Term v

Term representation for outcome expressions.

Different terms may represent equal outcomes.

Constructors

Var v
Variable, to be substituted by a more concrete outcome.

Never
Outcome that never finishes.

Wait0
Succeed immediately. Equivalent to Wait 0, but with a straight line as graphical representation.

Wait Rational
Succeed after waiting for a fixed amount of time.

Seq [Term v]
Sequential composition.

Last [Term v]
Parallel composition, last to finish.

First [Term v]
Parallel composiiton, last to finish.

isVar :: Term v -> Bool

Check whether a Term is a Var.

isSeq :: Term v -> Bool

Check whether a Term is a Seq.

isLast :: Term v -> Bool

Check whether a Term is a Last.

isFirst :: Term v -> Bool

Check whether a Term is a First.

isParallel :: Term v -> Bool

Check whether a Term is a parallel operation, i.e. Last or First.

maxParallel :: Term v -> Int

Maximal number of outcomes that run in parallel.

everywhere :: (Term v -> Term v) -> Term v -> Term v

Apply a transformation everywhere; bottom-up.

See also Scrap your boilerplate .

isNormalizedAssoc :: Term v -> Bool

Predicate that defines the "associative normal form" for a Term.

Specifically, a Term is said to be in "associative normal form" if the arguments to the constructors Seq, Last, and First

are nonempty lists, and

their list elements do not have an outermost constructor of the same kind, i.e. Seq [Seq …, First …] is not allowed because one of the list elements for a Seq constructor is also a Seq constructor.

normalizeAssoc :: Term v -> Term v

Normalize a term to "associative normal form".