DeltaQ.Methods

Description

This module collects common methods for constructing DeltaQ and analyzing them.

Slack / Hazard

meetsRequirement :: DeltaQ o => o -> (Duration o, Probability o) -> SlackOrHazard o

Compute "slack or hazard" with respect to a given reference point.

data SlackOrHazard o

The "slack or hazard" represents the distance between a reference point in (time, probability) space and a given DeltaQ.

Slack represents the case where the DeltaQ meets the performance requirements set by the reference point.

Hazard represents the case where the DeltaQ fails to meet the performance requirements set by the reference point.

Both cases include information of how far the reference point is away.

Constructors

Slack (Duration o) (Probability o)
We have some slack. Specifically, we have Duration at the same probability as the reference, and Probability at the same duration as the reference.

Hazard (Eventually (Duration o)) (Probability o)
We fail to meet the reference point. Specifically, we overshoot by Duration at the same probability as the reference, and by Probability at the same duration as the reference.

isSlack :: SlackOrHazard o -> Bool

Test whether the given SlackOrHazard is Slack.

isHazard :: SlackOrHazard o -> Bool

Test whether the given SlackOrHazard is Hazard.

Timeouts

retryOverlapN :: Int -> Duration DQ -> DQ -> DQ

Retry an outcome n times, take the first instance that succeeds.

More specifically, retryOverlapN n dt o models the completion times of a race of n independent copies of the outcome o against each other, succeeding when any of the copies first succeeds.

Copy number 0 <= j < n is started after waiting a time j * dt.

If no copy succeeds within total time n * dt, the whole outcome fails.

retryOverlap :: [Duration DQ] -> DQ -> DQ

Retry an outcome multiple times, take the first instance that succeeds.

More specifically, retryOverlap dts o models the completion times of a race of length dts independent copies of the outcome o against each other, succeeding when any of the copies first succeeds.

Each copy is started after a timeout in the event that the previous copies have not finished yet; the list dts gives the successive timeouts.

If no copy succeeds within total time sum dts, the whole outcome fails.