DeltaQ.PiecewisePolynomial

Description

Durations a is a probability distribution of completion time suing numeric type a. This type represents a mixed discrete / continuous probability distribution where the continuous part is represented in terms of piecewise polynomials.

Documentation

newtype Durations time

Probability distribution of durations.

Constructors

Durations

Fields

unDurations :: IRV time