DeltaQ.PiecewisePolynomial

Description

DQ is a probability distribution of completion time using the numeric type Rational. This type represents a mixed discrete / continuous probability distribution where the continuous part is represented in terms of piecewise polynomials.

Documentation

data DQ

Probability distribution of durations.

distribution :: DQ -> Piecewise (Poly Rational)

Get the distribution function as piecewise function of polynomials.

fromPositiveMeasure :: Measure Rational -> Maybe DQ

Interpret a finite, signed Measure as a probability distribution.

In order to admit an interpretation as probability, the measure needs to be positive, if that is not the case, the function returns Nothing.

Note: Due to rare numerical issues, the check isPositive can be false negative. Work around using unsafeFromPositiveMeasure.

unsafeFromPositiveMeasure :: Measure Rational -> DQ

Interpret a finite, positive Measure as a probability distribution.

/The precondition that the measure satisfies isPositive is not checked!/

Internal

complexity :: DQ -> Int

Size of the representation of a probability distribution, i.e. number of pieces of the piecewise function and degrees of the polynomials.

This quantity is relevant to stating and analyzing the asymptotic time complexity of operations.